Standardize and reduce error handling by using `?`
Googling for "rust Result ?" does not yield a good result, but suffice
to say that `?` propagates errors up the stack so you don't have to
manually specify each type of error manually.

This necessitates implementing the `From` trait for a couple of error
types so they can be converted automatically into the `Error` enum we
have created for ourselves.